requirements.txt
pandas
psycopg2-binary
sqlalchemy
pycountry
requests
apache-airflow[postgres]==2.9.1
apache-airflow-providers-postgres==5.1.0 # Für lokale Entwicklung